feedbackbot/requirements.txt
aiofiles==23.2.1
aiogram==3.13.1
aiogram-media-group==0.5.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.1.3
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.0
httpx==0.27.0
idna==3.4
Jinja2==3.1.4
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.4
orjson==3.10.6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.0
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.0
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
uvicorn==0.23.2
uvloop==0.20.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
